fix: errors

+++ controllers/cards.js
@@ -35,6 +35,9 @@ module.exports.deleteCard = (req, res) => {
       if (!card) {
         return res.status(404).send({ message: 'Карточка с указанным _id не найдена' });
       }
+      if (!card._id) {
+        return res.status(400).send({ message: 'Получение карточки с некорректным id' });
+      }
       Card.findByIdAndRemove(req.params.cardId)
         .then(() => res.status(200).send({ message: 'Карточка удалена' }));
     })
